Split body-kind probe tests into own file to fix SwiftLint
The 5 probe-emission tests added to `BridgeGeneratorTraitsTests` pushed
the host struct body to 1088 lines, over SwiftLint's 1000-line
`type_body_length` ceiling — the lint job on PR #163 failed with
exactly this violation. Moving them to their own focused
`BodyKindProbeEmissionTests` suite restores the host struct under the
limit and gives the body-kind probe a focused home distinct from the
trait-injection tests it was sharing space with.

Also swap the magic `1` in `bodyKindProbeReturnsOneForSwiftUI` for
`BodyKind.swiftUI.rawCode` so the contract change is enforced at one
more site.
